Extract extensions out of utility
This should make it a bit easier to find the expanded classes
