+++ lib/azure/core/utility.rb
@@ -117,74 +117,3 @@ module Azure
   end
 end
 
-class String
-  { reset:  0,
-    bold:  1,
-    dark:  2,
-    underline:  4,
-    blink:  5,
-    orange:  6,
-    negative:  7,
-    black: 30,
-    red: 31,
-    green: 32,
-    yellow: 33,
-    blue: 34,
-    magenta: 35,
-    cyan: 36,
-    white: 37,
-  }.each do |key, value|
-    define_method key do
-      "\e[#{value}m" + self + "\e[0m"
-    end
-  end
-end
-
-# Code validate private/public IP acceptable ranges.
-class IPAddr
-  PRIVATE_RANGES = [
-    IPAddr.new('10.0.0.0/8'),
-    IPAddr.new('172.16.0.0/12'),
-    IPAddr.new('192.168.0.0/16')
-  ]
-
-  def private?
-    return false unless self.ipv4?
-    PRIVATE_RANGES.each do |ipr|
-      return true if ipr.include?(self)
-    end
-    false
-  end
-
-  def public?
-    !private?
-  end
-
-  class << self
-    def validate_ip_and_prefix(ip, cidr)
-      if cidr.to_s.empty?
-        raise "Cidr is missing for IP '#{ip}'."
-      elsif valid?(ip)
-        raise "Ip address '#{ip}' is invalid."
-      elsif !IPAddr.new(ip).private?
-        raise "Ip Address #{ip} must be private."
-      end
-    end
-
-    def validate_address_space(ip)
-      if ip.split('/').size != 2
-        raise "Cidr is invalid for IP #{ip}."
-      elsif valid?(ip)
-        raise "Address space '#{ip}' is invalid."
-      end
-    end
-
-    def address_prefix(ip, cidr)
-      ip + '/' + cidr.to_s
-    end
-
-    def valid?(ip)
-      (IPAddr.new(ip) rescue nil).nil?
-    end
-  end
-end
