+++ lib/core_ext/string.rb
@@ -1,0 +1,23 @@
+class String
+  { reset:  0,
+    bold:  1,
+    dark:  2,
+    underline:  4,
+    blink:  5,
+    orange:  6,
+    negative:  7,
+    black: 30,
+    red: 31,
+    green: 32,
+    yellow: 33,
+    blue: 34,
+    magenta: 35,
+    cyan: 36,
+    white: 37,
+  }.each do |key, value|
+    define_method key do
+      "\e[#{value}m" + self + "\e[0m"
+    end
+  end
+end
+
